Fix the build with podman 4.x
Latest podman entered Debian testing. This new version doesn't depend on
iptables anymore, so as long as we need iptables-legacy, we need to pass
it to the apt command line explicitly.

Second, due to some changes in podman internals, building the extra
images now requires the network, otherwise it fails with cryptic error
messages:

    STEP 2/3: RUN echo "deb http://http.kali.org/kali kali-bleeding-edge main contrib non-free" >/etc/apt/sources.list.d/kali-bleeding-edge.list
    error running container: did not get container start message from parent: EOF
    Error: building at STEP "RUN echo "deb http://http.kali.org/kali kali-bleeding-edge main contrib non-free" >/etc/apt/sources.list.d/kali-bleeding-edge.list":
      netavark: No such file or directory (os error 2)

If I'm lucky, this commit is enough to fix it all.