Exclude /dev when packing the rootfs
As fas as I know, /dev is populated when the container is started, so
there's really no need to pack it in the rootfs archive.

So let's exclude it. This is what debuerreotype does for Debian images.

As a benefit, no need for root to export the container image, eg:

  mkdir chroot
  docker export <<uid>> | tar -C chroot -xf -

Additionally, it fixes a failure encountered while using the images in
the GitLab CI pipeline: there's a line that creates /dev/urandom (in the
piuparts job), and it fails because /dev/urandom is already there.