Use 'buildah build' instead of 'podman build'
Due to bug https://github.com/containers/buildah/issues/4427, which is
solved in buildah, but not yet in podman (it would need a rebuild of the
podman package in Debian).

This commit can be reverted when podman is fixed. Or, the opposite, we
could use buildah in place of podman everywhere, although I don't see
any reason to do so, and my understanding so far is that both commands
are equivalent, for the sake of building images.